- explicit 64-bit size for exported integer types, eg. sequence, identifier - format strings updated for newly explicit size of integers - Cygwin: receiver will ignore packets not from the expected sender - support for ARM, MIPS & PowerPC CPUs - comments in code for portability issues - new DEBUG_TRAFFIC conditional #define - tidy up configure.ac, dump.c & status.h

+++ src/subscriber.c
@@ -122,7 +122,7 @@ static void *stats_func(thread_handle thr)
 	if (!delim_pos)
 	    strncpy(alias, "unknown", sizeof(alias));
 	else {
-	    int sz = delim_pos - storage_desc;
+	    size_t sz = delim_pos - storage_desc;
 	    if (sz >= sizeof(alias))
 		sz = sizeof(alias) - 1;
